+++ app/lib/pages/recall/empty_state.dart
@@ -1,0 +1,105 @@
+import 'package:flutter/material.dart';
+import 'package:tdesign_flutter/tdesign_flutter.dart';
+import '../../configs/app_config.dart';
+import '../webgl_viewer.dart';
+
+/// 空状态组件
+class EmptyState extends StatelessWidget {
+  final bool isDark;
+  final Color textColor;
+  final Color hintTextColor;
+  final TDThemeData theme;
+  final Color darkCard;
+  final Color darkBorder;
+
+  const EmptyState({
+    super.key,
+    required this.isDark,
+    required this.textColor,
+    required this.hintTextColor,
+    required this.theme,
+    required this.darkCard,
+    required this.darkBorder,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    final iconColor = isDark
+        ? const Color(0xFFEEEEEE)
+        : const Color(0xFF333333);
+
+    return Center(
+      child: Container(
+        width: MediaQuery.of(context).size.width * 0.85,
+        padding: const EdgeInsets.symmetric(vertical: 64, horizontal: 24),
+        decoration: BoxDecoration(
+          color: isDark ? darkCard : theme.whiteColor1.withAlpha(200),
+          borderRadius: BorderRadius.circular(32.0),
+          border: Border.all(
+            color: isDark ? darkBorder : theme.whiteColor1,
+            width: 1,
+          ),
+          boxShadow: [
+            BoxShadow(
+              color: Colors.black.withAlpha(20),
+              blurRadius: 20,
+              spreadRadius: 5,
+            ),
+          ],
+        ),
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            TDImage(
+              assetUrl: 'assets/sprites/empty_state.png',
+              width: 120,
+              height: 120,
+              errorWidget: Icon(
+                TDIcons.time_filled,
+                size: 80,
+                color: iconColor,
+              ),
+            ),
+            const SizedBox(height: 24),
+            TDText(
+              textLocalize("home_page"),
+              font: theme.fontTitleLarge,
+              textColor: textColor,
+              fontWeight: FontWeight.w600,
+            ),
+            const SizedBox(height: 8),
+            TDText(
+              textLocalize("recall_empty_title"),
+              font: theme.fontBodyMedium,
+              textColor: hintTextColor,
+            ),
+            const SizedBox(height: 40),
+            TDButton(
+              text: textLocalize("recall_open_demo"),
+              iconWidget: Icon(
+                TDIcons.view_module,
+                color: Colors.white,
+                size: 20,
+              ),
+              type: TDButtonType.fill,
+              theme: TDButtonTheme.primary,
+              shape: TDButtonShape.round,
+              size: TDButtonSize.large,
+              onTap: () {
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                    builder: (context) => WebGLViewerPage(
+                      sceneId: textLocalize("recall_demo_title"),
+                    ),
+                  ),
+                );
+              },
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
